Remove iframe from starter page templates preview
